pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompts-tool"
version = "0.1.0"
description = "一个本地安装的 CLI 工具，用于检索、填充变量、生成最终 Prompt"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "streamlit>=1.28.0",
    "pyperclip>=1.8.2",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
prompts = "prompts_tool.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/prompts-tool"
Repository = "https://github.com/yourusername/prompts-tool"
Issues = "https://github.com/yourusername/prompts-tool/issues"

[tool.hatch.build.targets.wheel]
packages = ["prompts_tool"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
